+++ cmd/compositing/composite.js
@@ -26,27 +26,64 @@
 // Strict mode.
 'use strict'
 
-// Standard lib.
-const path = require('path')
-
 // Local modules.
 const constants = require('../../lib/constants')
 const queue = require('../../lib/queue')
 
+// Helpers.
+function getValueAt (arrayLike, index) {
+  if (Array.isArray(arrayLike)) {
+    return arrayLike[Math.min(index, arrayLike.length - 1)]
+  }
+  return arrayLike
+}
+
+function hasOwnProperty (obj, prop) {
+  return obj != null && Object.prototype.hasOwnProperty.call(obj, prop)
+}
+
 // Configure.
+const positionals = {
+  images: {
+    desc: 'Path to an image file',
+    normalize: true,
+    type: 'string'
+  }
+}
+
 const options = {
+  animated: {
+    desc: 'Read all frames/pages of an animated image',
+    type: 'boolean'
+  },
   blend: {
     choices: constants.BLEND,
-    default: 'over',
+    defaultDescription: 'over',
     desc: 'How to blend this image with the image below',
-    nargs: 1,
     type: 'string'
   },
   create: {
     desc: 'Describes a blank overlay to be created',
-    defaultDescription: 'width, height, channels, background',
-    nargs: 4,
-    type: 'array'
+    hidden: true,
+    type: 'boolean'
+  },
+  'create.background': {
+    desc: 'Background of the blank overlay to be created, parsed by the color module',
+    type: 'string'
+  },
+  'create.channels': {
+    choices: [3, 4],
+    default: 3,
+    desc: 'Number of channels of the blank overlay to be created',
+    type: 'number'
+  },
+  'create.height': {
+    desc: 'Height of the blank overlay to be created',
+    type: 'number'
+  },
+  'create.width': {
+    desc: 'Width of the blank overlay to be created',
+    type: 'number'
   },
   density: {
     desc: 'Number representing the DPI for vector images',
@@ -54,29 +91,27 @@ const options = {
     nargs: 1,
     type: 'number'
   },
-  gravity: {
-    choices: constants.GRAVITY,
-    default: 'centre',
-    desc: 'Gravity at which to place the overlay',
-    nargs: 1,
+  failOn: {
+    choices: constants.FAIL_ON,
+    defaultDescription: 'warning',
+    desc: 'Level of sensitivity to invalid images',
     type: 'string'
   },
-  image: {
-    coerce: path.normalize, // Positional arguments need manual normalization.
-    desc: 'Path to an image file',
-    normalize: true,
+  gravity: {
+    choices: constants.GRAVITY,
+    defaultDescription: 'centre',
+    desc: 'Gravity at which to place the overlay',
     type: 'string'
+  },
+  left: {
+    desc: 'The pixel offset from the left edge',
+    implies: 'top',
+    type: 'number'
   },
   limitInputPixels: {
     defaultDescription: 0x3FFF * 0x3FFF,
     desc: 'Do not process input images where the number of pixels (width x height) exceeds this limit',
-    nargs: 1,
     type: 'number'
-  },
-  offset: {
-    desc: 'The pixel offset from the top and left edges',
-    nargs: 2,
-    type: 'array'
   },
   premultiplied: {
     desc: 'Avoid premultiplying the image',
@@ -85,6 +120,11 @@ const options = {
   tile: {
     desc: 'Repeat the overlay image across the entire image with the given gravity',
     type: 'boolean'
+  },
+  top: {
+    desc: 'The pixel offset from the top edge',
+    implies: 'left',
+    type: 'number'
   }
 }
 
@@ -94,7 +134,16 @@ const builder = (yargs) => {
   return yargs
     .strict()
     .example('$0 composite ./input.png --gravity southeast', 'The output will be the input composited with ./input.png with SE gravity')
+    .example('$0 composite ./layer1.png --gravity northwest ./layer2.png --gravity southeast', 'The output will be the input composited with ./layer1.png with NW gravity and ./layer2.png with SE gravity')
+    .example('$0 composite ./overlay.png --tile --blend saturate', 'The output will be the input composited with ./overlay.png saturated over the entire image')
     .epilog('For more information on available options, please visit http://sharp.pixelplumbing.com/api-composite#composite')
+    .positional('images', positionals.images)
+    .check(argv => {
+      if ((!argv.images || argv.images.length === 0) && !hasOwnProperty(argv.create, 'width')) {
+        throw new Error('Expected at least one of images positional or create option')
+      }
+      return true
+    })
     .options(options)
     .global(optionNames, false)
     .group(optionNames, 'Command Options')
@@ -102,30 +151,54 @@ const builder = (yargs) => {
 
 // Command handler.
 const handler = (args) => {
-  const [top, left] = args.offset || []
-  const [width, height, channels, background] = args.create || []
+  if ((hasOwnProperty(args.create, 'width') && !hasOwnProperty(args.create, 'height')) ||
+   (hasOwnProperty(args.create, 'height') && !hasOwnProperty(args.create, 'width'))) {
+    throw new Error('Expected both of create.width and create.height options')
+  }
+
+  const inputs = []
+  if (hasOwnProperty(args.create, 'width')) {
+    const { channels, background, width, height } = args.create
+    const length = Math.max(width.length ?? 0, height.length ?? 0)
+    if (length > 0) { // Width or height (or both) is an array.
+      inputs.push(...Array.from({ length }, (_, idx) => idx).map(idx => ({
+        create: {
+          width: getValueAt(width, idx),
+          height: getValueAt(height, idx),
+          channels: getValueAt(channels, idx),
+          background: getValueAt(background, idx)
+        }
+      })))
+    } else { // Both width and height are numbers.
+      inputs.push({ create: args.create })
+    }
+  }
+  if (args.images) inputs.push(...args.images)
 
   // @see http://sharp.pixelplumbing.com/api-composite#composite
   return queue.push(['composite', (sharp) => {
-    return sharp.composite([{
-      input: args.image,
-      blend: args.blend,
-      create: args.create && { width, height, channels, background },
-      density: args.density,
-      gravity: args.gravity,
-      left,
-      limitInputPixels: args.limitInputPixels,
-      premultiplied: args.premultiplied,
-      tile: args.tile,
-      top
-    }])
+    return sharp.composite(
+      inputs.map((input, idx) => ({
+        animated: getValueAt(args.animated, idx),
+        blend: getValueAt(args.blend, idx),
+        density: getValueAt(args.density, idx),
+        failOn: getValueAt(args.failOn, idx),
+        gravity: getValueAt(args.gravity, idx),
+        input,
+        left: getValueAt(args.left, idx),
+        limitInputPixels: getValueAt(args.limitInputPixels, idx),
+        premultiplied: getValueAt(args.premultiplied, idx),
+        tile: getValueAt(args.tile, idx),
+        top: getValueAt(args.top, idx)
+      }))
+    )
   }])
 }
 
 // Exports.
 module.exports = {
-  command: 'composite <image>',
-  describe: 'Composite image over the processed (resized, extracted etc.) image',
+  command: 'composite [images..]',
+  describe: 'Composite image(s) over the processed (resized, extracted etc.) image',
   builder,
   handler
 }